requirements.txt
cassandra-driver
elasticsearch
blist
ipython
